Initial (non-working/tested) binary output for some tables.
The following classes have a new toBinary() function, designed to return a representation of table data in binary according to TrueType/OpenType specification. It's incomplete and they merely serve as an outline for future development:

- tables.head.head
- tables.hhea.hhea
- tables.vhea.vhea

Some of these table values have also been renamed to match what they are called in the TrueType/OpenType standards and have better comments.