Most class names converted to PascalCase
Generally to make the code a lot more readable and clearer. (Also this 
conforms to Python's writing style guidelines)

Whoops.

This isnt the case however when it comes to lowercase font table names, 
they're staying the same. (This is because font table names are case 
sensitive. I'll try to think of a way around it in the future)